% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsf.R
\name{bicTuning}
\alias{bicTuning}
\title{Tuning Parameter Selection via the Bayesian Information Criterion}
\usage{
bicTuning(y, result)
}
\arguments{
\item{y}{n by D matrix consisting of the data.}

\item{result}{A \code{gsf} object.}
}
\value{
A list containing the selected tuning parameter and 
        corresponding estimates, as well as a summary of the
        computed RBIC values, log-likelihood values, and corresponding
        orders of the estimates.
}
\description{
Bayesian Information Criterion (BIC) for tuning parameter selection.
}
\details{
The BIC selects the best tuning parameter out of the ones used in a
 \code{gsf} object by minimizing the following function

\deqn{\textrm{BIC}(\lambda) = -2 l_n(\hat{\mathbf{\Psi}}_{\lambda}) + 
\textrm{df}(\hat{\mathbf{\Psi}}_{\lambda}) \log n}

where \eqn{l_n} is the log-likelihood function, and \eqn{
\hat{\mathbf{\Psi}}_{\lambda}} is the set of estimated 
parameters \code{theta} and \code{pii} corresponding to the tuning parameter \eqn{\lambda}.
\eqn{\textrm{df}(\hat{\mathbf{\Psi}}_{\lambda})} denotes the degrees
of freedom of the estimates.
}
\examples{
 data(faithful)
 out <- normalLocOrder(faithful, K=12, 
                       lambdas=c(0.1, 0.25, 0.5, 0.75, 1.0), penalty="MCP", a=2)
 plot(out)
 bicTuning(faithful, out)

}
\references{
Manole, T., Khalili, A. 2019. "Estimating the Number of Components in Finite Mixture Models 
 via the Group-Sort-Fuse Procedure".
}
