% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{enzyme}
\alias{enzyme}
\title{Enzyme Dataset.}
\format{A vector with 245 elements.}
\source{
Bechtel, Y. C., Bonaiti-Pellie, C., Poisson, N., Magnette, 
 J., & Bechtel, P. R. 1993. "A population and family study 
 N-acetyltransferase using caffeine urinary metabolites". 
 Clinical Pharmacology & Therapeutics: 54(2), 134-141.
}
\usage{
enzyme
}
\description{
A dataset containing the distribution of the N-acetyltransferase
enzyme in the blood of 245 individuals.
}
\keyword{datasets}
