% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsf.R
\name{exponentialOrder}
\alias{exponentialOrder}
\title{Estimate the Number of Components in a Finite Mixture of Exponential Distributions.}
\usage{
exponentialOrder(y, lambdas, K = NULL, theta, ...)
}
\arguments{
\item{y}{Vector consisting of the data.}

\item{lambdas}{Vector of tuning parameter values.}

\item{K}{Upper bound on the true number of components. 
If \code{K} is \code{NULL}, at least one of \code{theta} and
\code{pii} must be non-\code{NULL}, and K is inferred from their
number of columns.}

\item{theta}{Vector of starting values for the Exponential distribution parameters, of length K.}

\item{...}{Additional control parameters. See the \strong{Details} section.}
}
\value{
An object with S3 classes \code{gsf} and \code{exponentialGsf}, 
        consisting of a list with the estimates produced for every tuning
        parameter in \code{lambdas}.
}
\description{
Estimate the order of a finite mixture of Exponential distributions.
}
\details{
The following is a list of additional control parameters.

\itemize{
  \item{\code{pii}}    {Vector of size K whose elements must sum to 1, consisting of 
                        the starting values for the mixing proportions.  
                        If \code{NULL}, it will be set to a discrete 
                        uniform distribution with K support points.}
  \item{\code{arbSigma}} {\code{TRUE} if the common variance-covariance matrix should
                     be estimated, and FALSE if it should stay fixed, and equal to
                     \code{sigma}.}
  \item{\code{penalty}} {Choice of penalty, which may be \code{"SCAD"}, \code{"MCP"} or 
                         \code{"ADAPTIVE-LASSO"}. Default is \code{"SCAD"}.}
  \item{\code{a}}        {Tuning parameter for the SCAD or MCP penalty. Default is \code{3.7}.}
  \item{\code{mcmcIter}} {Number of iterations for the starting value algorithm described 
                     in the details below). This value is ignored when \code{theta} 
                     is not \code{NULL}. }
  \item{\code{uBound}}   {Upper bound on the tuning parameter of the PGD algorithm.}
  \item{\code{C}}       {Tuning parameter for penalizing the mixing proportions.}
  \item{\code{convMem}}  {Convergence criterion for the modified EM algorithm.}
  \item{\code{convPgd}}    {Convergence criterion for the proximal gradient descent algorithm.}
  \item{\code{maxMem}}   {Maximum number of iterations of the modified EM algorithm.}
  \item{\code{maxPgd}}   {Maximum number of iterations of the proximal gradient descent algorithm.}
  \item{\code{verbose}}  {If \code{TRUE}, print updates while the function is running.}
}
}
\references{
Manole, T., Khalili, A. 2019. "Estimating the Number of Components in Finite Mixture Models 
 via the Group-Sort-Fuse Procedure".
}
