% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsf.R
\name{multinomialOrder}
\alias{multinomialOrder}
\title{Estimate the Number of Components in a Multinomial Mixture Model.}
\usage{
multinomialOrder(y, lambdas, K = NULL, ...)
}
\arguments{
\item{y}{n by D matrix consisting of the data, where n is the sample size 
and D is the number of categories.
The rows of \code{y} must sum to a constant value, 
taken to be the number of trials.}

\item{lambdas}{Vector of tuning parameter values.}

\item{K}{Upper bound on the true number of components. 
If \code{K} is \code{NULL}, at least one of the control parameters \code{theta} and
\code{pii} must be non-\code{NULL}, and K is inferred from their
number of columns.}

\item{...}{Additional control parameters. See the \strong{Details} section.}
}
\value{
An object with S3 classes \code{gsf} and \code{multinomialGsf}, 
        consisting of a list with the estimates produced for every tuning
        parameter in \code{lambdas}.
}
\description{
Estimate the order of a finite mixture of multinomial models with fixed and 
known number of trials.
}
\details{
The following is a list of additional control parameters.

\itemize{
  \item{\code{theta}}    {D by K matrix of starting values where each column is the vector
                     of multinomial probabilities for one mixture component.              
                     The columns of \code{theta} should therefore 
                     sum to 1. If \code{theta=NULL}, the starting values are
                     chosen using the MCMC algorithm described by Grenier (2016).} 
  \item{\code{pii}}      {Vector of size K whose elements must sum to 1, consisting of 
                     the starting values for the mixing proportions.  
                     If \code{NULL}, it will be set to a discrete 
                     uniform distribution with K support points.}
  \item{\code{penalty}} {Choice of penalty, which may be \code{"SCAD"}, \code{"MCP"}, 
                         \code{"SCAD-LLA"}, \code{"MCP-LLA"} or \code{"ADAPTIVE-LASSO"}. 
                         Default is \code{"SCAD"}.}
  \item{\code{mcmcIter}} {Number of iterations for the starting value algorithm described 
                          by Grenier (2016).}
  \item{\code{uBound}}   {Upper bound on the tuning parameter of the proximal gradient algorithm.}
  \item{\code{C}}       {Tuning parameter for penalizing the mixing proportions.}
  \item{\code{a}}        {Tuning parameter for the SCAD or MCP penalty. Default is \code{3.7}.}
  \item{\code{convMem}}  {Convergence criterion for the modified EM algorithm.}
  \item{\code{convPgd}}    {Convergence criterion for the proximal gradient descent algorithm.}
  \item{\code{maxMem}}   {Maximum number of iterations of the Modified EM algorithm.}
  \item{\code{maxPgd}}   {Maximum number of iterations of the proximal gradient descent algorithm.}
  \item{\code{verbose}}  {If \code{TRUE}, print updates while the function is running.}
}
}
\examples{
 require(MM)
 data(pollen)
 set.seed(1)
 out <- multinomialOrder(pollen, K=12, lambdas=seq(0.1, 1.6, 0.2))
 tuning <- bicTuning(pollen, out)
 plot(out, eta=TRUE, gg=FALSE, opt=tuning$result$lambda)
}
\references{
Manole, T., Khalili, A. 2019. "Estimating the Number of Components in Finite Mixture Models 
 via the Group-Sort-Fuse Procedure".

 Grenier, I. (2016) Bayesian Model Selection for Deep Exponential Families. 
 M.Sc. dissertation, McGill University Libraries.
}
