% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{notices}
\alias{notices}
\title{Number of Death Notices Dataset.}
\format{A dataframe with 1095 columns.}
\source{
Schilling, W. 1947. "A Frequency Distribution Represented as the 
        Sum of Two Poisson Distributions". Journal of the American Statistical 
        Association. 42 (239): 407-424.
}
\usage{
notices
}
\description{
A dataset containing the number of death notices of women eighty years of age
and over, appearing in the Times of London on each day for the three consecutive
years 1910-1912.
}
\keyword{datasets}
