% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsf.R
\name{plot.gsf}
\alias{plot.gsf}
\title{Plot the coefficient paths of an \code{gsf} object.}
\usage{
\method{plot}{gsf}(x, gg = FALSE, eta = TRUE, vlines = TRUE,
  points = FALSE, opt = NULL, sqrtScale = FALSE, ...)
}
\arguments{
\item{x}{An \code{gsf} object.}

\item{gg}{A ggplot2 figure is produced if \code{TRUE}, and a Base R figure is produced otherwise.}

\item{eta}{If \code{TRUE}, a coefficient plot for the distance between consecutive coefficient is produced, and a 
plot with all coefficients is produced otherwise.}

\item{vlines}{If \code{TRUE}, vertical dashed lines are plotted whenever the number of components decreases, 
\code{FALSE} otherwise.}

\item{points}{If \code{TRUE}, points are plotted along the lines, and are not plotted otherwise.}

\item{opt}{Optimal tuning parameter value. If not \code{NULL}, a vertical red line is drawn at this point.}

\item{sqrtScale}{\code{TRUE} if lambda values are to be plotted on the square root scale, \code{FALSE} otherwise.}

\item{...}{Additional Base R plotting parameters.}
}
\value{
Returns one or more plots depending on the value of \code{eta}.
}
\description{
Plot the coefficient paths of an \code{gsf} object.
}
\examples{
 data(faithful)
 out <- normalLocOrder(faithful, K=12, 
                       lambdas=c(0.1, 0.25, 0.5, 0.75, 1.0), penalty="MCP", a=2)
 plot(out)
 bicTuning(faithful, out)

}
\references{
Manole, T., Khalili, A. 2019. "Estimating the Number of Components in Finite Mixture Models 
 via the Group-Sort-Fuse Procedure".
}
