% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsf.R
\name{poissonOrder}
\alias{poissonOrder}
\title{Estimate the Number of Components in a Poisson Finite Mixture Model.}
\usage{
poissonOrder(y, lambdas, K = NULL, ...)
}
\arguments{
\item{y}{Vector consisting of the data.}

\item{lambdas}{Vector of tuning parameter values.}

\item{K}{Upper bound on the true number of components. 
If \code{K} is \code{NULL}, at least one of \code{theta} and
\code{pii} must be non-\code{NULL}, and K is inferred from their
number of columns.}

\item{...}{Additional control parameters. See the \strong{Details} section.}
}
\value{
An object with S3 classes \code{gsf} and \code{poissonGsf}, 
        consisting of a list with the estimates produced for every tuning
        parameter in \code{lambdas}.
}
\description{
Estimate the order of a finite mixture of Poisson distributions.
}
\details{
The following is a list of additional control parameters.

\itemize{
  \item{\code{theta}}  {Vector of starting values for the Poisson parameters, of length K.
                        If \code{NULL},  the starting values are chosen to be a discrete 
                        uniform distribution on the 100(k- 1/2)/K \% sample quantiles.}
  \item{\code{pii}}    {Vector of size K whose elements must sum to 1, consisting of 
                        the starting values for the mixing proportions.  
                        If \code{NULL}, it will be set to a discrete 
                        uniform distribution with K support points.}
  \item{\code{penalty}} {Choice of penalty, which may be \code{"SCAD"}, \code{"MCP"}, 
                         \code{"SCAD-LLA"}, \code{"MCP-LLA"} or \code{"ADAPTIVE-LASSO"}. 
                         Default is \code{"SCAD"}.}
  \item{\code{a}}        {Tuning parameter for the SCAD or MCP penalty. Default is \code{3.7}.}
  \item{\code{uBound}}   {Upper bound on the tuning parameter of the proximal gradient descent algorithm.}
  \item{\code{C}}       {Tuning parameter for penalizing the mixing proportions.}
  \item{\code{convMem}}  {Convergence criterion for the modified EM algorithm.}
  \item{\code{convPgd}}    {Convergence criterion for the proximal gradient descent algorithm.}
  \item{\code{maxMem}}   {Maximum number of iterations of the modified EM algorithm.}
  \item{\code{maxPgd}}   {Maximum number of iterations of the proximal gradient descent algorithm.}
  \item{\code{verbose}}  {If \code{TRUE}, print updates while the function is running.}
}
}
\examples{
 data(notices)

 # Run the GSF with the MCP penalty.
 set.seed(1)
 out <- poissonOrder(notices, lambdas=c(.001, .005, .01, .1, .5, 1, 2, 5), 
                     K=12, uBound=0.01, penalty="MCP")
 # Requires ggplot2.
 plot(out, eta=FALSE)

 # Run the GSF with the Adaptive Lasso penalty.
 set.seed(1)
 out <- poissonOrder(notices, lambdas=c(.001, .01, .5, 1, 2), 
                     K=12, penalty="ADAPTIVE-LASSO")
 plot(out, eta=FALSE)

 # Run the GSF with the SCAD penalty.
 set.seed(1)
 out <- poissonOrder(notices, lambdas=c(.00005, .001, .005, .01, .1, .5, 1, 2, 5), 
                     K=12, uBound=0.01, penalty="SCAD")
 plot(out, eta=FALSE)

 # Select a tuning parameter using the BIC. 
 bicTuning(notices, out)
}
\references{
Manole, T., Khalili, A. 2019. "Estimating the Number of Components in Finite Mixture Models 
 via the Group-Sort-Fuse Procedure".
}
