% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsf.R
\name{bicTuning}
\alias{bicTuning}
\title{Tuning Parameter Selection via the Bayesian Information Criterion}
\usage{
bicTuning(y, result)
}
\arguments{
\item{y}{n by D matrix consisting of the data.}

\item{result}{A \code{gsf} object.}
}
\value{
A list containing the selected tuning parameter and 
        corresponding estimates, as well as a summary of the
        computed RBIC values, log-likelihood values, and corresponding
        orders of the estimates.
}
\description{
Bayesian Information Criterion (BIC) for tuning parameter selection.
}
\details{
The BIC selects the best tuning parameter out of the ones used in a
 \code{gsf} object by minimizing the following function

\deqn{\textrm{BIC}(\lambda) = -2 l_n(\hat{\mathbf{\Psi}}_{\lambda}) + 
\textrm{df}(\hat{\mathbf{\Psi}}_{\lambda}) \log n}

where \eqn{l_n} is the log-likelihood function, and \eqn{
\hat{\mathbf{\Psi}}_{\lambda}} is the set of estimated 
parameters \code{theta} and \code{pii} corresponding to the tuning parameter \eqn{\lambda}.
\eqn{\textrm{df}(\hat{\mathbf{\Psi}}_{\lambda})} denotes the degrees
of freedom of the estimates.
}
\examples{
 require(MM)
 data(pollen)
 set.seed(1)
 out <- multinomialOrder(pollen, K=12, lambdas=seq(0.1, 1.6, 0.2))
 tuning <- bicTuning(pollen, out)
 plot(out, eta=TRUE, gg=FALSE, opt=tuning$result$lambda)
}
